Don't try to mount artifacts during sapling-builder build
docker apparently isn't able to mount volumes during builds like podman is, so
instead identify the most recent fbthrift build from the host script and use a
COPY directive to get it into the sapling build container.

+++ scripts/build-sapling.sh
@@ -5,7 +5,17 @@ set -e
 SCRIPTS=$(cd "$(dirname "$0")" && pwd)
 . "$SCRIPTS/lib"
 
+
+LATEST_FBTHRIFT="$(printf '%s\n' ./artifacts/x86_64/fbthrift-*.el10.x86_64.tar.xz | sort | tail -n1)"
+if [ ! -f "$LATEST_FBTHRIFT" ]; then
+	echo "No fbthrift build found in ./artifacts/ ! Run scripts/build-fbthrift.sh first" >&2
+	exit 1
+fi
+
+echo "Using fbthrift from ${LATEST_FBTHRIFT}"
+ln "$LATEST_FBTHRIFT" sapling-builder/files/fbthrift.tar.xz
+
 BUILDER_IMAGE_ID="$(mktemp)"
-"$DOCKER" build --iidfile="$BUILDER_IMAGE_ID" -v "$PWD/artifacts:/artifacts:ro,z" ./sapling-builder
+"$DOCKER" build --iidfile="$BUILDER_IMAGE_ID" ./sapling-builder
 "$DOCKER" run -v ./artifacts:/artifacts:z "$(cat "$BUILDER_IMAGE_ID")"
 rm -f "$BUILDER_IMAGE_ID"
